+++ system/platform-core/src/main/java/org/platformlambda/core/system/WorkerHandler.java
@@ -450,5 +450,4 @@ public class WorkerHandler {
         }
         return response;
     }
-
 }