+++ system/platform-core/src/main/java/org/platformlambda/core/websocket/server/MinimalistHttpHandler.java
@@ -245,5 +245,4 @@ public class MinimalistHttpHandler implements Handler<HttpServerRequest> {
         }
         return null;
     }
-
 }