+++ system/platform-core/src/main/java/org/platformlambda/core/websocket/server/WsRequestHandler.java
@@ -228,7 +228,5 @@ public class WsRequestHandler implements Handler<ServerWebSocket> {
             }
             return null;
         }
-
     }
-
 }