+++ CHANGELOG.md
@@ -8,6 +8,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.0.27, 11/27/2024
+
+### Added
+
+1. Support for Mono/Flux return type for KotlinLambdaFunction
+2. Implemented Websocket handshake handler to adjust to API changes in vertx 4.5.11
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
 ## Version 4.0.26, 11/26/2024
 
 ### Added